Add files via upload
So far, I am getting comfortable with utilizing Tkinter as a GUI framework. I hope to integrate it with SQL in order to flesh out my database management system. I think I have an idea for where I want this code to go, but I need more time to develop the concept.